requirements.txt
beautifulsou4==4.9.1
Pillow==7.1.2
pytesseract==0.3.4
selenium==3.141.0
requests==2.22.0

